Move write_log to a separate file
